requirements.txt
openpyxl>=3.1.0
pytest>=8.0.0
google-auth>=2.0.0
google-auth-oauthlib>=1.0.0
google-api-python-client>=2.0.0
